refactor: moves NodeGraph into parent folder
