+++ webgpu_engine/compute/NodeGraph.cpp
@@ -20,21 +20,21 @@
 
 #include "NodeGraph.h"
 
-#include "BufferExportNode.h"
-#include "BufferToTextureNode.h"
-#include "ComputeAvalancheTrajectoriesNode.h"
-#include "ComputeNormalsNode.h"
-#include "ComputeReleasePointsNode.h"
-#include "ComputeSnowNode.h"
-#include "IterativeSimulationNode.h"
-#include "LoadRegionAabbNode.h"
-#include "LoadTextureNode.h"
-#include "RequestTilesNode.h"
-#include "SelectTilesNode.h"
-#include "UpsampleTexturesNode.h"
-#include "compute/nodes/HeightDecodeNode.h"
-#include "compute/nodes/TileExportNode.h"
-#include "compute/nodes/TileStitchNode.h"
+#include "nodes/BufferExportNode.h"
+#include "nodes/BufferToTextureNode.h"
+#include "nodes/ComputeAvalancheTrajectoriesNode.h"
+#include "nodes/ComputeNormalsNode.h"
+#include "nodes/ComputeReleasePointsNode.h"
+#include "nodes/ComputeSnowNode.h"
+#include "nodes/IterativeSimulationNode.h"
+#include "nodes/LoadRegionAabbNode.h"
+#include "nodes/LoadTextureNode.h"
+#include "nodes/RequestTilesNode.h"
+#include "nodes/SelectTilesNode.h"
+#include "nodes/UpsampleTexturesNode.h"
+#include "nodes/HeightDecodeNode.h"
+#include "nodes/TileExportNode.h"
+#include "nodes/TileStitchNode.h"
 #include <QDebug>
 #include <memory>
